fixed tsc issues
