+++ insights-ui/src/components/home-page/FromTheBlog.tsx
@@ -1,8 +1,3 @@
-// app/components/FromTheBlog.tsx
-import fs from 'fs';
-import path from 'path';
-import matter from 'gray-matter';
-
 interface PostData {
   id: string;
   title: string;
@@ -10,44 +5,11 @@ interface PostData {
   description: string;
   date: string;
   datetime: string;
-  category: { title: string; href: string };
+  category: { title: string; href: string }[];
   image: string;
 }
 
-// ✅ Helper function to read all MDX files from the blogs directory
-async function getPostsData(): Promise<PostData[]> {
-  const postsDirectory = path.join(process.cwd(), 'blogs');
-  const fileNames = fs.readdirSync(postsDirectory);
-
-  const posts = fileNames.map((fileName) => {
-    const slug = fileName.replace(/\.mdx?$/, '');
-    const filePath = path.join(postsDirectory, fileName);
-    const fileContents = fs.readFileSync(filePath, 'utf8');
-    const { data } = matter(fileContents);
-
-    return {
-      id: slug, // using the slug as a unique id
-      title: data.title || 'Untitled Post',
-      href: `/blogs/${slug}`,
-      description: data.description || 'No description available.',
-      date: data.date || 'Unknown Date',
-      datetime: data.datetime || data.date || 'Unknown Date',
-      category: {
-        title: data.category.title || 'General',
-        href: '#',
-      },
-      image: data.image || '/images/default-thumbnail.jpg',
-    };
-  });
-
-  return posts;
-}
-
-// ✅ Async Server Component using Next.js App Router
-export default async function FromTheBlog() {
-  // Fetch posts data from MDX files
-  const posts = await getPostsData();
-
+export default function FromTheBlog({ posts }: { posts: PostData[] }) {
   return (
     <div className="bg-gray-800 py-24 sm:py-32">
       <div className="mx-auto max-w-7xl px-6 lg:px-8">
@@ -65,9 +27,11 @@ export default async function FromTheBlog() {
                 <time dateTime={post.datetime} className="text-gray-500">
                   {post.date}
                 </time>
-                <a href={post.category.href} className="relative z-10 rounded-full bg-gray-700 px-3 py-1.5 font-medium text-gray-300 hover:bg-gray-600">
-                  {post.category.title}
-                </a>
+                {post.category.map((cat, index) => (
+                  <a key={index} href={cat.href} className="relative z-10 rounded-full bg-gray-700 px-3 py-1.5 font-medium text-gray-300 hover:bg-gray-600">
+                    {cat.title}
+                  </a>
+                ))}
               </div>
               <div className="group relative">
                 <h3 className="mt-3 text-lg font-semibold text-white group-hover:text-indigo-400">
